+++ ui/src/pages/admin/bookings/[id].tsx
@@ -16,22 +16,23 @@ import Space from "@/types/Space";
 import User from "@/types/User";
 import Location from "@/types/Location";
 import Booking from "@/types/Booking";
-import Ajax from "@/util/Ajax";
 import UserPreference from "@/types/UserPreference";
 import Formatting from "@/util/Formatting";
 import FullLayout from "@/components/FullLayout";
 
 import DateUtil from "@/util/DateUtil";
 import DateTimePicker from "@/components/DateTimePicker";
-import RuntimeConfig from "@/components/RuntimeConfig";
 import RendererUtils from "@/util/RendererUtils";
 import ProfilePicture from "@/components/ProfilePicture";
 import Search, { SearchOptions } from "@/types/Search";
+import AjaxError from "@/util/AjaxError";
+import ErrorText from "@/types/ErrorText";
 
 interface State {
   loading: boolean;
   saved: boolean;
   error: boolean;
+  errorText: string;
   wasCreated: boolean;
   goBack: boolean;
   enter: Date;
@@ -52,8 +53,8 @@ interface State {
   canSave: boolean;
   canEdit: boolean;
   prefEnterTime: number;
-  prefWorkdayStart: number;
-  prefWorkdayEnd: number;
+  prefWorkdayStart: string;
+  prefWorkdayEnd: string;
   prefWorkdays: number[];
   prefLocationId: string;
   selfEmail: string;
@@ -118,8 +119,8 @@ class EditBooking extends React.Component<Props, State> {
       canSave: false,
       canEdit: false,
       prefEnterTime: 0,
-      prefWorkdayStart: 0,
-      prefWorkdayEnd: 0,
+      prefWorkdayStart: "09:00",
+      prefWorkdayEnd: "17:00",
       prefWorkdays: [],
       prefLocationId: "",
       selfEmail: "",
@@ -272,9 +273,9 @@ class EditBooking extends React.Component<Props, State> {
               if (s.name === UserPreference.PREF_ENTER_TIME)
                 state.prefEnterTime = window.parseInt(s.value);
               if (s.name === UserPreference.PREF_WORKDAY_START)
-                state.prefWorkdayStart = window.parseInt(s.value);
+                state.prefWorkdayStart = DateUtil.parseTimeString(s.value);
               if (s.name === UserPreference.PREF_WORKDAY_END)
-                state.prefWorkdayEnd = window.parseInt(s.value);
+                state.prefWorkdayEnd = DateUtil.parseTimeString(s.value);
               if (s.name === UserPreference.PREF_WORKDAYS)
                 state.prefWorkdays = s.value
                   .split(",")
@@ -292,8 +293,8 @@ class EditBooking extends React.Component<Props, State> {
               state.prefBuddyBookedColor = s.value;
           });
           if (self.dailyBasisBooking) {
-            state.prefWorkdayStart = 0;
-            state.prefWorkdayEnd = 23;
+            state.prefWorkdayStart = "00:00";
+            state.prefWorkdayEnd = "23:59";
           }
           self.setState(
             {
@@ -337,10 +338,9 @@ class EditBooking extends React.Component<Props, State> {
       let leave = new Date();
       leave = this.state.leave;
       leave.setHours(23, 59, 59, 0);
-
       this.setState({
-        enter: enter,
-        leave: leave,
+        enter,
+        leave,
       });
     } else {
       const enter = this.state.enter;
@@ -378,9 +378,16 @@ class EditBooking extends React.Component<Props, State> {
             selectedUserEmail: user,
           });
         })
-        .catch(() => {
+        .catch((e) => {
+          let code: number = 0;
+          if (e instanceof AjaxError) {
+            code = e.appErrorCode;
+          }
           this.setState({
             error: true,
+            errorText: code
+              ? ErrorText.getTextForAppCode(code, this.props.t)
+              : this.props.t("errorSave"),
             saved: false,
             wasCreated: true,
           });
@@ -406,9 +413,16 @@ class EditBooking extends React.Component<Props, State> {
             wasCreated: false,
           });
         })
-        .catch(() => {
+        .catch((e) => {
+          let code: number = 0;
+          if (e instanceof AjaxError) {
+            code = e.appErrorCode;
+          }
           this.setState({
             error: true,
+            errorText: code
+              ? ErrorText.getTextForAppCode(code, this.props.t)
+              : this.props.t("errorSave"),
             saved: false,
             wasCreated: false,
           });
@@ -531,8 +545,8 @@ class EditBooking extends React.Component<Props, State> {
       }
       this.setState(
         {
-          enter: enter,
-          leave: leave,
+          enter,
+          leave,
           isDisabledLocation: false,
           isDisabledSpace: true,
         },
@@ -721,7 +735,7 @@ class EditBooking extends React.Component<Props, State> {
         <Alert variant="danger">{this.props.t(this.state.canSearchHint)}</Alert>
       );
     } else if (this.state.error) {
-      hint = <Alert variant="danger">{this.props.t("errorSave")}</Alert>;
+      hint = <Alert variant="danger">{this.state.errorText}</Alert>;
     }
 
     const buttonDelete = (