+++ ui/src/pages/preferences.tsx
@@ -6,6 +6,8 @@ import { IoLinkOutline } from "react-icons/io5";
 import NavBar from "@/components/NavBar";
 import withReadyRouter from "@/components/withReadyRouter";
 import RuntimeConfig from "@/components/RuntimeConfig";
+import DateTimePicker from "@/components/DateTimePicker";
+import DateUtil from "@/util/DateUtil";
 import { TranslationFunc, withTranslation } from "@/components/withTranslation";
 import Ajax from "@/util/Ajax";
 import UserPreference from "@/types/UserPreference";
@@ -32,8 +34,8 @@ interface State {
   saved: boolean;
   error: boolean;
   enterTime: number;
-  workdayStart: number;
-  workdayEnd: number;
+  workdayStart: string;
+  workdayEnd: string;
   workdays: number[];
   booked: string;
   notBooked: string;
@@ -95,8 +97,8 @@ class Preferences extends React.Component<Props, State> {
       saved: false,
       error: false,
       enterTime: 0,
-      workdayStart: 0,
-      workdayEnd: 0,
+      workdayStart: "09:00",
+      workdayEnd: "17:00",
       workdays: [],
       booked: COLOR_BOOKED,
       notBooked: COLOR_NOT_BOOKED,
@@ -159,11 +161,11 @@ class Preferences extends React.Component<Props, State> {
       if (typeof window !== "undefined") {
         if (s.name === UserPreference.PREF_ENTER_TIME)
           state.enterTime = window.parseInt(s.value);
-        if (s.name === UserPreference.PREF_WORKDAY_START)
-          state.workdayStart = window.parseInt(s.value);
-        if (s.name === UserPreference.PREF_WORKDAY_END)
-          state.workdayEnd = window.parseInt(s.value);
       }
+      if (s.name === UserPreference.PREF_WORKDAY_START)
+        state.workdayStart = DateUtil.parseTimeString(s.value) ?? "09:00";
+      if (s.name === UserPreference.PREF_WORKDAY_END)
+        state.workdayEnd = DateUtil.parseTimeString(s.value) ?? "17:00";
       if (s.name === UserPreference.PREF_WORKDAYS) {
         state.workdays = s.value
           .split(",")
@@ -229,11 +231,11 @@ class Preferences extends React.Component<Props, State> {
       ),
       new UserPreference(
         UserPreference.PREF_WORKDAY_START,
-        this.state.workdayStart.toString(),
+        this.state.workdayStart,
       ),
       new UserPreference(
         UserPreference.PREF_WORKDAY_END,
-        this.state.workdayEnd.toString(),
+        this.state.workdayEnd,
       ),
       new UserPreference(
         UserPreference.PREF_WORKDAYS,
@@ -347,7 +349,7 @@ class Preferences extends React.Component<Props, State> {
   };
 
   onWorkdaysChange = (workdays: number[]) => {
-    this.setState({ workdays });
+    this.setState({ workdays: [...workdays].sort((a, b) => a - b) });
   };
 
   connectCalDav = async () => {
@@ -580,23 +582,37 @@ class Preferences extends React.Component<Props, State> {
                 <Form.Label htmlFor="workdayStart">
                   {this.props.t("workingHours")}
                 </Form.Label>
-                <div>
-                  <Form.Control
-                    type="number"
-                    id="workdayStart"
-                    value={this.state.workdayStart}
-                    onChange={(e: any) =>
-                      this.setState({
-                        workdayStart:
-                          typeof window !== "undefined"
-                            ? window.parseInt(e.target.value)
-                            : 0,
-                      })
-                    }
-                    min="0"
-                    max="23"
-                    style={{ display: "inline", width: "40%" }}
-                  />
+                <div className="d-flex align-items-center">
+                  <div style={{ width: "40%" }}>
+                    <DateTimePicker
+                      id="workdayStart"
+                      noCalendar={true}
+                      enableTime={true}
+                      value={DateUtil.getTodayTimeFromTimeString(
+                        this.state.workdayStart,
+                      )}
+                      maxDate={DateUtil.getTodayTimeFromMinutes(23 * 60 - 1)}
+                      onChange={(value: Date) => {
+                        const workdayStart = DateUtil.formatTimeString(value);
+                        const startMinutes =
+                          DateUtil.timeStringToMinutes(workdayStart);
+                        const endMinutes = DateUtil.timeStringToMinutes(
+                          this.state.workdayEnd,
+                        );
+                        this.setState({
+                          workdayStart: workdayStart,
+                          workdayEnd:
+                            endMinutes >= startMinutes + 60
+                              ? this.state.workdayEnd
+                              : DateUtil.formatTimeString(
+                                  DateUtil.getTodayTimeFromMinutes(
+                                    startMinutes + 60,
+                                  ),
+                                ),
+                        });
+                      }}
+                    />
+                  </div>
                   <span
                     style={{
                       width: "20%",
@@ -608,20 +624,44 @@ class Preferences extends React.Component<Props, State> {
                       {this.props.t("to").toString()}
                     </Form.Label>
                   </span>
-                  <Form.Control
-                    type="number"
-                    id="workdayEnd"
-                    value={this.state.workdayEnd}
-                    onChange={(e: any) =>
-                      this.setState({ workdayEnd: e.target.value })
-                    }
-                    min={this.state.workdayStart + 1}
-                    max="23"
-                    style={{ display: "inline", width: "40%" }}
-                  />
+                  <div style={{ width: "40%" }}>
+                    <DateTimePicker
+                      id="workdayEnd"
+                      noCalendar={true}
+                      enableTime={true}
+                      value={DateUtil.getTodayTimeFromTimeString(
+                        this.state.workdayEnd,
+                      )}
+                      minDate={DateUtil.getTodayTimeFromMinutes(
+                        DateUtil.timeStringToMinutes(this.state.workdayStart) +
+                          60,
+                      )}
+                      onChange={(value: Date) => {
+                        const workdayEnd = DateUtil.formatTimeString(value);
+                        const endMinutes =
+                          DateUtil.timeStringToMinutes(workdayEnd);
+                        const startMinutes = DateUtil.timeStringToMinutes(
+                          this.state.workdayStart,
+                        );
+                        this.setState({
+                          workdayEnd: workdayEnd,
+                          workdayStart:
+                            startMinutes <= endMinutes - 60
+                              ? this.state.workdayStart
+                              : DateUtil.formatTimeString(
+                                  DateUtil.getTodayTimeFromMinutes(
+                                    endMinutes - 60,
+                                  ),
+                                ),
+                        });
+                      }}
+                    />
+                  </div>
                 </div>
                 {!RuntimeConfig.INFOS.dailyBasisBooking &&
-                  this.state.workdayEnd - this.state.workdayStart >
+                  (DateUtil.timeStringToMinutes(this.state.workdayEnd) -
+                    DateUtil.timeStringToMinutes(this.state.workdayStart)) /
+                    60 >
                     RuntimeConfig.INFOS.maxBookingDurationHours && (
                     <Form.Text muted>
                       {this.props.t("workingHoursHintExceedsMaxDuration", {